Support `include` and `subninja`.
We do this with a 'fileReference' event, rather than a separate
'include' and 'subninja' event. The reason is that a client will want to
handle both anyway, and the logic for handling them is so similar that
it makes sense for a client to do it with a single callback.